Trying to save the position and size of the window before closing.
Right now, the position always seems to save with a QPoint difference of +8 and +30,
i.e., the position gets saved as being 8 more pixels to the right, and 30 pixels more towards the bottom,
even if I haven't moved the window from its original position.
I think I need to write an onCloseEvent and get the position from there.