- In PlayerWindow::onMediaPlayerStatusChanged():     + Added prompt for user to delete file if the media player tries to play an invalid file.
- In PlayerWindow::initializeLibraryModels():
    + Changed the editStrategy of PlayerWindow's librarySource model to
QSqlTableModel::OnFieldChange so that the user doesn't have to manually
press Enter after editing the data of a cell, they can simply click away
and the changes will still be written to the media database. Before, if editting a
cell and user clicked away from it, the data would remain in cell
but not actually submit to the underlying media database; therefore when
restarting the Music Player, the previous changes to the cell were not
displayed.